fix some type problems

+++ include/constants.hpp
@@ -2,9 +2,9 @@
 
 namespace defaults {
 	size_t VOCAB_SIZE = 30000;
-	size_t HIDDEN_DIM = 1024;
-	size_t LAYERS = 2;
-	size_t INPUT_DIM = 256;
+	unsigned int HIDDEN_DIM = 1024;
+	unsigned int LAYERS = 2;
+	unsigned int INPUT_DIM = 256;
 	size_t EPOCHS = 5;
 	size_t BATCH_SIZE = 20;
 }
